+++ _doc/sphinxdoc/source/conf.py
@@ -13,8 +13,8 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.split(__file__)[0])))
 local_template = os.path.join(os.path.abspath(
     os.path.dirname(__file__)), "phdoc_templates")
 
-set_sphinx_variables(__file__, "mlinsights", "Xavier Dupré", 2022,
-                     "pydata_sphinx_theme", pydata_sphinx_theme.get_html_theme_path(),
+set_sphinx_variables(__file__, "mlinsights", "Xavier Dupré", 2021,
+                     "pydata_sphinx_theme", ['_static'],
                      locals(), extlinks=dict(
                          issue=('https://github.com/sdpython/mlinsights/issues/%s', 'issue')),
                      title="mlinsights", book=True)