+++ _unittests/ut_search_rank/test_LONG_search_images_keras.py
@@ -28,9 +28,10 @@ class TestSearchPredictionsImagesKeras(ExtTestCase):
         with redirect_stderr(StringIO()):
             try:
                 from keras.applications.mobilenet import MobileNet  # pylint: disable=E0401,E0611
-            except (SyntaxError, ModuleNotFoundError, AttributeError) as e:
+            except (SyntaxError, ModuleNotFoundError, AttributeError,
+                    ImportError) as e:
                 warnings.warn(
-                    "Issue with tensorflow or keras: {0}".format(e))
+                    f"Issue with tensorflow or keras: {e}")
                 return
             from keras.preprocessing.image import ImageDataGenerator  # pylint: disable=E0401,E0611
             from keras.preprocessing.image import img_to_array, load_img  # pylint: disable=E0401,E0611