+++ mlinsights/search_rank/search_engine_predictions_images.py
@@ -56,8 +56,7 @@ class SearchEnginePredictionImages(SearchEnginePredictions):
                     "".format(type(iter_images)))
             if iter_images.batch_size != 1:
                 raise ValueError(  # pragma: no cover
-                    "batch_size must be 1 not {0}".format(
-                        iter_images.batch_size))
+                    f"batch_size must be 1 not {iter_images.batch_size}")
             self.iter_images_ = iter_images
             if n is None:
                 n = len(iter_images)
@@ -66,7 +65,7 @@ class SearchEnginePredictionImages(SearchEnginePredictions):
                     "Iterator does not iterate on images but numpy arrays (not implemented).")
         else:
             raise TypeError(  # pragma: no cover
-                "Unexpected data type {0}.".format(type(data)))
+                f"Unexpected data type {type(data)}.")
 
         def get_current_index(flow):
             "get current index"
@@ -87,11 +86,11 @@ class SearchEnginePredictionImages(SearchEnginePredictions):
                 im, name = acc(i, it)
                 if not isinstance(name, str):
                     raise TypeError(  # pragma: no cover
-                        "name should be a string, not {0}".format(type(name)))
+                        f"name should be a string, not {type(name)}")
                 yield im[0], dict(name=name, i=i)
                 if fLOG and i % 10000 == 0:
                     fLOG(
-                        '[SearchEnginePredictionImages.fit] i={}/{} - {}'.format(i, n, name))
+                        f'[SearchEnginePredictionImages.fit] i={i}/{n} - {name}')
         super()._prepare_fit(data=iterator_feature_meta(), transform=transform)
 
     def fit(self, iter_images, n=None, fLOG=None):  # pylint: disable=W0237
@@ -127,22 +126,21 @@ class SearchEnginePredictionImages(SearchEnginePredictions):
                 X = from_numpy(iter_images[numpy.newaxis, :, :, :])
                 return super().kneighbors(X, n_neighbors=n_neighbors)
             raise RuntimeError(  # pragma: no cover
-                "Unknown module '{0}'.".format(self.module_))
+                f"Unknown module '{self.module_}'.")
         elif "keras" in str(iter_images):  # pragma: no cover
             if self.module_ != "keras":
                 raise RuntimeError(  # pragma: no cover
-                    "Keras object but {0} was used to train the KNN.".format(self.module_))
+                    f"Keras object but {self.module_} was used to train the KNN.")
             # We delay the import as keras backend is not necessarily installed.
             # keras, it expects an iterator.
             from keras.preprocessing.image import Iterator  # pylint: disable=E0401,C0415,E0611
             from keras_preprocessing.image import DirectoryIterator, NumpyArrayIterator  # pylint: disable=E0401,C0415,E0611
             if not isinstance(iter_images, (Iterator, DirectoryIterator, NumpyArrayIterator)):
                 raise NotImplementedError(  # pragma: no cover
-                    "iter_images must be a keras Iterator. No option implemented for type {0}.".format(type(iter_images)))
+                    f"iter_images must be a keras Iterator. No option implemented for type {type(iter_images)}.")
             if iter_images.batch_size != 1:
                 raise ValueError(  # pragma: no cover
-                    "batch_size must be 1 not {0}".format(
-                        iter_images.batch_size))
+                    f"batch_size must be 1 not {iter_images.batch_size}")
             for img in iter_images:
                 X = img[0]
                 break
@@ -150,7 +148,7 @@ class SearchEnginePredictionImages(SearchEnginePredictions):
         elif "torch" in str(type(iter_images)):
             if self.module_ != "torch":
                 raise RuntimeError(  # pragma: no cover
-                    "Torch object but {0} was used to train the KNN.".format(self.module_))
+                    f"Torch object but {self.module_} was used to train the KNN.")
             # torch: it expects a tensor
             X = iter_images
             return super().kneighbors(X, n_neighbors=n_neighbors)